requirements.txt
pandas>=1.5.0
PyYAML>=6.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
matplotlib>=3.6.3
sendgrid==6.11.0
